Move server module
