+++ python/src/aac/cli.py
@@ -9,7 +9,7 @@ from pluggy import PluginManager
 
 from aac import __version__
 from aac.AacCommand import AacCommand, AacCommandArgument
-from aac.lang.server import start_lsp
+from aac.lang.lsp.server import start_lsp
 from aac.plugins.plugin_manager import get_plugin_manager
 from aac.plugins.plugin_execution import PluginExecutionResult, plugin_result
 from aac.spec.core import get_aac_spec_as_yaml
